ver. 0.015

+++ src/app/ReadFile.java
@@ -3,5 +3,5 @@ package app;
 import java.io.IOException;
 
 public interface ReadFile {
-    String readLinesFromFileByID(int ID) throws IOException;
+    String readLinesFromFileByID(long ID) throws IOException;
 }
